Preserve final newline in Jinja templates
As described in the pocoo-libs thread "Bug in jinja2.lexer",
https://groups.google.com/forum/?fromgroups=#!topic/pocoo-libs/6DylMqq1voI
Jinja normalizes newlines in a manner that loses a newline at the end
of a file.

Add a check for the trailing newline in a template to be rendered by
JinjaFormatter and add an extra newline to preserve the original in
the output.